data_processing/requirements.txt
google-cloud-storage
google-cloud-pubsub
requests
pandas
pyyaml
trino
